requirements.txt
Django==4.2.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
Pillow
python-dotenv==1.0.0
